job-pipeline/pyproject.toml
[project]
name = "job-pipeline"
version = "0.1.0"
description = "Personal job application pipeline - Stage 1: Job Aggregator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-jobspy>=1.1.0",
    "psycopg2-binary>=2.9.9",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "tabulate>=0.9.0",
    "schedule>=1.2.0",
    "pandas>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["discovery", "dashboard"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
addopts = "-v --tb=short"
